% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewTask-f2dbJob.R, R/addTask-f2dbJob.R
\docType{methods}
\name{addNewTask-method}
\alias{addNewTask-method}
\alias{addNewTask}
\alias{addNewTask,f2dbJob-method}
\alias{addTask-method}
\alias{addTask}
\alias{addTask,f2dbJob,f2dbTask-method}
\alias{addTask,f2dbJob,list-method}
\title{addTask, addNewTask}
\usage{
addNewTask(job, ...)

\S4method{addNewTask}{f2dbJob}(job, name, taskFunction, ..., inputName, itemName)

addTask(job, task, followNext, ...)

\S4method{addTask}{f2dbJob,f2dbTask}(job, task, followNext = FALSE)

\S4method{addTask}{f2dbJob,list}(job, task, followNext = FALSE)
}
\arguments{
\item{job}{An \code{f2dbJob} to add a task to.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Parameters passed to the task
function.}

\item{name}{Task name.}

\item{taskFunction}{Function to call when running the task.}

\item{inputName}{By default, task input is passed to the task function as an
unnamed first parameter. If \code{inputName} is provided, task input will be
passed using \code{inputName}.}

\item{itemName}{By default, the current batch item is not passed to the task
function. If \code{itemName} is provided, the batch item will be passed using
\code{itemName}.}

\item{task}{An existing \code{f2dbTask} to add.}

\item{followNext}{Recursively adds \code{nextTask}}
}
\value{
\code{NULL}, invisibly.
}
\description{
Add a \code{f2dbTask} object to the end of the task list and update the \code{nextTask}
slot of the next-to-last task to point to the newly added task.
}
\seealso{
Other f2dbJob: 
\code{\link{f2dbJob}},
\code{\link{f2dbJob-class}},
\code{\link{jobInput-set-method}}
}
\concept{f2dbJob}
