% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendTask-f2dbJob.R
\docType{methods}
\name{appendTask-method}
\alias{appendTask-method}
\alias{appendTask}
\alias{appendTask,f2dbJob,f2dbTask-method}
\title{appendTask}
\usage{
appendTask(job, task)

\S4method{appendTask}{f2dbJob,f2dbTask}(job, task)
}
\arguments{
\item{job}{The \code{f2dbJob} to add a job to.}

\item{task}{The \code{f2dbTask} to add.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Adds an existing \code{f2dbTask} object to the end of the task list and updates
the \code{nextTask} slot of the next-to-last task to point to the newly added task.
}
\seealso{
Other f2dbJob: 
\code{\link{f2dbJob}},
\code{\link{f2dbJob-class}},
\code{\link{jobInput-set-method}}
}
\concept{f2dbJob}
