% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-f2dbBatch.R
\docType{class}
\name{f2dbBatch-class}
\alias{f2dbBatch-class}
\title{f2dbBatch class}
\description{
A container for jobs.
}
\section{Slots}{

\describe{
\item{\code{name}}{Batch name.}

\item{\code{jobList}}{List of jobs}
}}

\seealso{
Other f2dbBatch: 
\code{\link{f2dbBatch}},
\code{\link{jobList-method}},
\code{\link{jobList-set-method}}

Other f2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbJob-class}},
\code{\link{f2dbListTask-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbBatch}
