% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-f2dbBatch.R
\name{f2dbBatch}
\alias{f2dbBatch}
\title{f2dbBatch}
\usage{
f2dbBatch(batchName)
}
\arguments{
\item{batchName}{Batch name.}
}
\value{
An \code{f2dbBatch} object.
}
\description{
\code{f2dbBatch} constructor.
}
\seealso{
Other f2dbBatch: 
\code{\link{f2dbBatch-class}},
\code{\link{f2dbRun,f2dbBatch-method}},
\code{\link{jobList-method}},
\code{\link{jobList-set-method}}
}
\concept{f2dbBatch}
