% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2dbEndTask.R
\docType{class}
\name{f2dbEndTask-class}
\alias{f2dbEndTask-class}
\title{f2dbEndTask class}
\description{
A \code{f2dbEndTask} object marks the end of a task list.
}
\details{
The \code{taskFunction} and \code{nextTask} slots of a \code{f2dbEndTask} object cannont be
modified.

\code{f2dbRun(f2dbEndTask)} will always return \code{TRUE}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbListTask-class}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}

Other f2dbEndTask: 
\code{\link{f2dbEndTask}}

Other f2db classes: 
\code{\link{f2dbBatch-class}},
\code{\link{f2dbJob-class}},
\code{\link{f2dbListTask-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbEndTask}
\concept{f2dbTask}
