% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2dbIterateTask.R
\docType{class}
\name{f2dbIterateTask-class}
\alias{f2dbIterateTask-class}
\title{f2dbIterateTask}
\description{
A \code{f2dbIterateTask} iterates over \code{taskInput}, running running
subsequent \code{taskFunction}s with each element as input.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}

Other f2dbIterateTask: 
\code{\link{f2dbIterateTask}},
\code{\link{f2dbRun,f2dbIterateTask-method}}

Other f2db classes: 
\code{\link{f2dbBatch-class}},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbJob-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbIterateTask}
\concept{f2dbTask}
