% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2dbIterateTask.R
\name{f2dbIterateTask}
\alias{f2dbIterateTask}
\title{f2dbIterateTask}
\usage{
f2dbIterateTask(taskName)
}
\arguments{
\item{taskName}{Task name.}
}
\value{
An \code{f2dbIterateTask} object.
}
\description{
\code{f2dbIterateTask} constructor.
}
\seealso{
Other f2dbIterateTask: 
\code{\link{f2dbIterateTask-class}},
\code{\link{f2dbRun,f2dbIterateTask-method}}
}
\concept{f2dbIterateTask}
