% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R
\docType{class}
\name{f2dbJob-class}
\alias{f2dbJob-class}
\title{f2dbJob class}
\description{
f2dbJob class
}
\section{Slots}{

\describe{
\item{\code{jobInput}}{Input supplied to the first task.}
}}

\seealso{
Other f2db classes: 
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbJob}
