% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R
\name{f2dbJob}
\alias{f2dbJob}
\title{f2dbJob}
\usage{
f2dbJob(name, input)
}
\arguments{
\item{name}{Job name. This should be unique within the job batch.}

\item{input}{Input supplied to the first task.}
}
\value{
An \code{f2dbJob} object.
}
\description{
\code{f2dbJob} constructor.
}
\seealso{
Other f2dbJob: 
\code{\link{appendTask-method}},
\code{\link{f2dbJob-class}},
\code{\link{jobInput-set-method}}
}
\concept{f2dbJob}
