% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2dbNullTask.R
\docType{class}
\name{f2dbNullTask-class}
\alias{f2dbNullTask-class}
\title{f2dbNullTask class}
\description{
A do-nothing task. This is the default nextTask for f2dbTask instances.
}
\section{Slots}{

\describe{
\item{\code{name}}{"NULLTASK"}
}}

