% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2dbNullTaskFunction.R
\docType{class}
\name{f2dbNullTaskFunction.class}
\alias{f2dbNullTaskFunction.class}
\title{f2dbNullTaskFunction class}
\description{
A do-nothing, pass-through task function. This is the default task function
when creating a new \code{f2dbTask} object.
}
\seealso{
Other file2dbTaskFunction: 
\code{\link{f2dbRun.f2dbTaskFunction}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}

Other file2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbObject.class}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask.class}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}
}
\concept{file2db classes}
\concept{file2dbTaskFunction}
