% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-f2dbObject.R
\docType{class}
\name{f2dbObject-class}
\alias{f2dbObject-class}
\title{f2dbObject class}
\description{
Base class from which all other f2db classes descend.
}
\section{Slots}{

\describe{
\item{\code{name}}{Object name.}
}}

\seealso{
Other f2dbObject: 
\code{\link{f2dbObject}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun-method}},
\code{\link{name-method}}

Other f2db classes: 
\code{\link{f2dbBatch-class}},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbIterateTask-class}},
\code{\link{f2dbJob-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbObject}
