% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-f2dbObject.R
\name{f2dbObject}
\alias{f2dbObject}
\title{f2dbObject}
\usage{
f2dbObject(name)
}
\arguments{
\item{name}{Object name.}
}
\value{
An \code{f2dbObject}.
}
\description{
\code{f2dbObject} constructor.
}
\seealso{
Other f2dbObject: 
\code{\link{f2dbObject-class}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun-method}},
\code{\link{name-method}}
}
\concept{f2dbObject}
