% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbBatch.R
\docType{methods}
\name{f2dbRun,f2dbBatch-method}
\alias{f2dbRun,f2dbBatch-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbBatch}(object)
}
\arguments{
\item{object}{An \code{f2dbBatch} to run.}
}
\value{
Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
There may be warning or informational messages even if success is indicated.
}
\description{
Run an \code{f2dbBatch}.
}
\details{
When called on an \code{f2dbBatch} object, \code{f2dbRun} will run the jobs in the
\code{jobList}.
}
\seealso{
Other f2dbBatch: 
\code{\link{f2dbBatch}},
\code{\link{f2dbBatch-class}},
\code{\link{jobList-method}},
\code{\link{jobList-set-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbIterateTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbBatch}
\concept{f2dbRun methods}
