% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbEndTask.R
\docType{methods}
\name{f2dbRun,f2dbEndTask-method}
\alias{f2dbRun,f2dbEndTask-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbEndTask}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbEndTask} object.}

\item{input}{\emph{Ignored}.}

\item{item}{\emph{Ignored}}
}
\value{
A list:
\item{success}{\code{TRUE}}
\item{output}{"{<}End of task list{>}"}
}
\description{
\code{f2dbRun(f2dbEndTask)} will always return \code{TRUE}.
}
\seealso{
Other f2dbEndTask: 
\code{\link{f2dbEndTask}},
\code{\link{f2dbEndTask-class}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbEndTask}
\concept{f2dbRun methods}
