% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbIterateTask.R
\docType{methods}
\name{f2dbRun,f2dbIterateTask-method}
\alias{f2dbRun,f2dbIterateTask-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbIterateTask}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbIterateTask} to run.}

\item{input}{Input to pass to the underlying task function.}

\item{item}{The job item being processed.}
}
\value{
Logical value indicating success (\code{TRUE}) or failure
(\code{FALSE}). \code{TRUE} is returned if both \code{taskFunction} and \code{nextTask} are
successful. Otherwise, \code{FALSE} is returned.
}
\description{
Run an \code{f2dbIterateTask}.
}
\details{
When called on an \code{f2dbIterateTask} object, \code{f2dbRun} will iterate over the
\code{taskInput}, running \code{nextTask} with \code{input} and \code{item} both set to the
\code{taskInput} element.
}
\seealso{
Other f2dbIterateTask: 
\code{\link{f2dbIterateTask}},
\code{\link{f2dbIterateTask-class}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbBatch-method}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbIterateTask}
\concept{f2dbRun methods}
