% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbJob.R
\docType{methods}
\name{f2dbRun,f2dbJob-method}
\alias{f2dbRun,f2dbJob-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbJob}(object)
}
\arguments{
\item{object}{An \code{f2dbJob} to run.}
}
\value{
A list:
\item{success}{Logical value indicating success (\code{TRUE}) or failure
(\code{FALSE}). There may be warning or informational messages even if success is
indicated.}
\item{object}{The type of object being run.}
\item{name}{The name of the object being run.}
\item{messages}{Captured error, warning, or informational messages.}
}
\description{
Run an \code{f2dbJob}.
}
\details{
When called on an \code{f2dbJob} object, \code{f2dbRun} will run the tasks in the
\code{taskList}. The first task is called using \code{jobInput} as both task \code{input}
and job \code{item}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbRun methods}
\concept{f2dbTask}
