% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbListTask.R
\docType{methods}
\name{f2dbRun,f2dbListTask-method}
\alias{f2dbRun,f2dbListTask-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbListTask}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbListTask} to run.}

\item{input}{Input to pass to the underlying task function.}

\item{item}{The job item being processed.}
}
\value{
Logical value indicating success (\code{TRUE}) or failure
(\code{FALSE}). \code{TRUE} is returned if both \code{taskFunction} and \code{nextTask} are
successful. Otherwise, \code{FALSE} is returned.
}
\description{
Run an \code{f2dbListTask}.
}
\details{
When called on an \code{f2dbListTask} object, \code{f2dbRun} will run \code{taskFunction}.
If \code{taskFunction} is successful, \code{f2dbRun} will then iterate over the output
of \code{taskFucntion}, running \code{nextTask} with \code{input} and \code{item} both set to the
output element.
}
\seealso{
Other f2dbListTask: 
\code{\link{f2dbListTask}},
\code{\link{f2dbListTask-class}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbBatch-method}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbListTask}
\concept{f2dbRun methods}
