% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbObject.R
\docType{methods}
\name{f2dbRun,f2dbObject-method}
\alias{f2dbRun,f2dbObject-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbObject}(object, ...)
}
\arguments{
\item{object}{An \code{f2dbObject}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Unused.}
}
\value{
A list:
\item{success}{\code{TRUE}.}
\item{object}{\code{f2dbObject}.}
\item{name}{The name of \code{object}.}
}
\description{
This is the default \code{f2dbRun} method for \code{f2dbObject}s.
}
\seealso{
Other f2dbObject: 
\code{\link{f2dbObject}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbRun-method}},
\code{\link{name-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbListTask-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbObject}
\concept{f2dbRun methods}
