% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-f2dbTaskFunction.R
\docType{methods}
\name{f2dbRun.f2dbTaskFunction}
\alias{f2dbRun.f2dbTaskFunction}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbTaskFunction}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbTaskFunction} object}

\item{input}{Task input to pass to the task function}

\item{item}{The current item being processed}
}
\value{
A list
\item{success}{Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
Success only indicates that it is safe to proceed; it does not mean there are
no warnings or errors.}
\item{output}{Run results, if any.}
}
\description{
Runs the given task function.
}
\details{
When called with an \code{\link[:f2dbNullTaskFunction.class]{f2dbNullTaskFunction}}
object, \code{f2dbRun} returns \code{list(success = TRUE, output = input)}.
}
\seealso{
Other f2dbTaskFunction: 
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}

Other f2dbRun methods: 
\code{\link{f2dbRun}()}
}
\concept{f2dbRun methods}
\concept{f2dbTaskFunction}
