% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-f2dbObject.R
\docType{methods}
\name{f2dbRun}
\alias{f2dbRun}
\title{f2dbRun}
\usage{
f2dbRun(object, ...)
}
\value{
A list
\item{success}{Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
Success only indicates that it is safe to proceed; it does not mean there are
no warnings or errors.}
\item{output}{Run results, if any.}
}
\description{
Generic method for running an f2dbObject.
}
\seealso{
Other f2dbRun methods: 
\code{\link{f2dbRun.f2dbTaskFunction}}
}
\concept{f2dbRun methods}
