% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-file2dbTask.R
\docType{class}
\name{f2dbTask-class}
\alias{f2dbTask-class}
\title{f2dbTask class}
\description{
A class for encapsulating an individual task.
}
\section{Slots}{

\describe{
\item{\code{name}}{Task name. This name normally be should be unique within the job.}

\item{\code{taskFunction}}{Function that performs the actual work of the task.}

\item{\code{nextTask}}{A f2dbTaskObject object that performs the next task in the
job sequence. nextTask takes the output of the current task as its input.}
}}

\seealso{
Other file2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}
}
\concept{file2db classes}
