% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-file2dbTask.R
\docType{class}
\name{f2dbTask.class}
\alias{f2dbTask.class}
\title{f2dbTask class}
\description{
A class for encapsulating an individual task.
}
\section{Slots}{

\describe{
\item{\code{name}}{Task name. This name normally be should be unique within the job.}

\item{\code{taskFunction}}{An \code{f2dbTaskFunction} to perform the actual work of the task.}

\item{\code{nextTask}}{An \code{f2dbTaskObject} object that performs the next task in the
job sequence. \code{nextTask} takes the output of the current task as its input.
This value is set when adding tasks to a job and should not be modified.}
}}

\seealso{
Other f2dbTask: 
\code{\link{f2dbTask}}

Other file2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbObject.class}},
\code{\link{f2dbTask}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}
}
\concept{f2dbTask}
\concept{file2db classes}
