% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R
\name{f2dbTask}
\alias{f2dbTask}
\title{f2dbTask}
\usage{
f2dbTask(name = NA, taskFunction, ..., inputName = NA, itemName = NA)
}
\arguments{
\item{name}{Task name. This name normally be should be unique within the job.}

\item{taskFunction}{Function to call when running the task}

\item{...}{Parameters to pass to the task function}

\item{inputName}{By default, task input is passed to the task function as an
unnamed first parameter. If \code{inputName} is provided, task input will be passed
using \code{inputName}.}

\item{itemName}{By default, the current batch item is not passed to the task
function. If \code{itemName} is provided, the batch item will be passed
using \code{itemName}.}
}
\description{
Creates a new \code{f2dbTask} object.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbTask.class}}

Other file2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbNullTaskFunction.class}},
\code{\link{f2dbObject.class}},
\code{\link{f2dbTask.class}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}
}
\concept{f2dbTask}
\concept{file2db classes}
