% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-f2dbTaskFunction.R
\docType{class}
\name{f2dbTaskFunction.class}
\alias{f2dbTaskFunction.class}
\title{f2dbTaskFunction class}
\arguments{
\item{env}{Environment in which the task function will be executed}
}
\description{
The \code{f2dbTaskFunction} class encapsulates the function that accomplishes the
actual work of the task. Along with the \code{\link[:f2dbRun.f2dbTaskFunction]{f2dbRun()}}
method, this class is responsible for:
\enumerate{
\item Calling the underlying task function
\item Capturing and logging any warnings or errors
\item Returning task function output
\item Returning whether or not the task executed successfully
}
}
\details{
Classes inheriting from \code{f2dbTaskFunction} should fulfill the responsibilities
listed above and implement the \code{f2dbRun} generic method.
}
\section{Slots}{

\describe{
\item{\code{taskFunction}}{A call to the underlying task function. \code{taskFunction} is
set by \code{\link[=f2dbTaskFunction]{f2dbTaskFunction()}} and should not be modified.}
}}

\seealso{
Other f2dbTaskFunction: 
\code{\link{f2dbRun.f2dbTaskFunction}},
\code{\link{f2dbTaskFunction}}

Other f2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbObject.class}},
\code{\link{f2dbTask.class}},
\code{\link{f2dbTaskFunction}}
}
\concept{f2db classes}
\concept{f2dbTaskFunction}
