% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-f2dbTaskFunction.R
\name{f2dbTaskFunction}
\alias{f2dbTaskFunction}
\title{f2dbTaskFunction}
\usage{
f2dbTaskFunction(
  taskFunction = NA,
  ...,
  inputName = NA,
  itemName = NA,
  env = rlang::caller_env()
)
}
\arguments{
\item{taskFunction}{Function to call when running the task}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Parameters passed to the task function}

\item{inputName}{By default, task input is passed to the task function as an
unnamed first parameter. If \code{inputName} is provided, task input will be passed
using \code{inputName}.}

\item{itemName}{By default, the current batch item is not passed to the task
function. If \code{itemName} is provided, the batch item will be passed
using \code{itemName}.}

\item{env}{Environment in which the task function will be executed}
}
\value{
An \code{f2dbTaskFunction} object
}
\description{
Creates a new \code{f2dbTaskFunction} object.
}
\seealso{
Other f2dbTaskFunction: 
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbTaskFunction-class}}

Other f2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbTaskFunction}
