% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2dbNullTask.R
\docType{class}
\name{file2dbNullTask-class}
\alias{file2dbNullTask-class}
\title{file2dbNullTask class}
\description{
A do-nothing task. This is the default nextTask for file2dbTask instances.
}
\section{Slots}{

\describe{
\item{\code{name}}{"NULLTASK"}
}}

