% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-file2dbTask.R
\docType{class}
\name{file2dbTask-class}
\alias{file2dbTask-class}
\title{file2dbTask class}
\description{
A virtual class for encapsulating an individual task.
}
\section{Slots}{

\describe{
\item{\code{name}}{Task name. This name should be unique within the job.}

\item{\code{taskFunction}}{Function that performs the actual work of the task.}

\item{\code{nextTask}}{A file2dbTaskObject object that performs the next task in the
job sequence. nextTask takes the output of the current task as its input.}
}}

