% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-file2dbTaskObject.R
\docType{class}
\name{file2dbTaskObject-class}
\alias{file2dbTaskObject-class}
\title{file2dbTaskObject class}
\description{
Base class from which all other file2db tasks should descend. This is a
virtual class.
}
