% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{get_dbconn}
\alias{get_dbconn}
\title{Get database connection}
\usage{
get_dbconn()
}
\value{
A DBIConnection object or NULL if no database connection has been set.
}
\description{
Gets the database connection previously set by \code{set_dbconn}.
}
\examples{
conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

set_dbconn(conn)

# Do things...

DBI::dbDisconnect(get_dbconn())

}
