% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2dbEndTask.R
\docType{methods}
\name{initialize,f2dbEndTask-method}
\alias{initialize,f2dbEndTask-method}
\title{initialize}
\usage{
\S4method{initialize}{f2dbEndTask}(.Object)
}
\arguments{
\item{.Object}{f2dbEndTask object}
}
\value{
Initialized f2dbEndTask object
}
\description{
Initializer function for f2dbEndTask objects. This initializer explicitly
avoids calling the parent f2dbTask initializer.
}
