% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2dbNullTask.R
\name{initialize,f2dbNullTask-method}
\alias{initialize,f2dbNullTask-method}
\title{Initialize function for f2dbNullTask objects}
\usage{
\S4method{initialize}{f2dbNullTask}(.Object)
}
\description{
This initializer explicitly avoids calling the parent f2dbTask initializer.
}
