% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R
\docType{methods}
\name{initialize,f2dbTask-method}
\alias{initialize,f2dbTask-method}
\title{initialize}
\usage{
\S4method{initialize}{f2dbTask}(.Object, ...)
}
\arguments{
\item{.Object}{f2dbTask object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Extra parameters}
}
\value{
Initialized f2dbTask object
}
\description{
Initializer function for f2dbTask objects.
}
