% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2dbNullTask.R
\name{initialize,file2dbNullTask-method}
\alias{initialize,file2dbNullTask-method}
\title{Initialize function for file2dbNullTask objects}
\usage{
\S4method{initialize}{file2dbNullTask}(.Object)
}
\description{
This initializer explicitly avoids calling the parent file2dbTask initializer.
}
