% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-file2dbTask.R
\name{initialize,file2dbTask-method}
\alias{initialize,file2dbTask-method}
\title{Default initialize function for file2dbTask objects}
\usage{
\S4method{initialize}{file2dbTask}(.Object)
}
\description{
Default initialize function for file2dbTask objects
}
