% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R
\docType{methods}
\name{jobInput-method}
\alias{jobInput-method}
\alias{jobInput}
\alias{jobInput,f2dbJob-method}
\title{jobInput}
\usage{
jobInput(object)

\S4method{jobInput}{f2dbJob}(object)
}
\arguments{
\item{object}{An \code{f2dbJob} object}
}
\value{
`jobInput
}
\description{
Returns job input to the first task
}
\concept{f2dbJob`}
