% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R
\docType{methods}
\name{jobInput-set-method}
\alias{jobInput-set-method}
\alias{jobInput<-}
\alias{jobInput<-,f2dbJob-method}
\title{jobInput<-}
\usage{
jobInput(object) <- value

\S4method{jobInput}{f2dbJob}(object) <- value
}
\arguments{
\item{object}{An \code{f2dbJob}}

\item{value}{\code{jobInput}}
}
\description{
Sets job input to the first task
}
\seealso{
Other f2dbJob: 
\code{\link{addNewTask-method}},
\code{\link{f2dbJob}},
\code{\link{f2dbJob-class}}
}
\concept{f2dbJob}
