% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-f2dbBatch.R
\docType{methods}
\name{jobList-method}
\alias{jobList-method}
\alias{jobList}
\alias{jobList,f2dbBatch-method}
\title{jobList}
\usage{
jobList(object)

\S4method{jobList}{f2dbBatch}(object)
}
\arguments{
\item{object}{An \code{f2dbBatch} object.}
}
\value{
A list of \code{f2dbJob}s.
}
\description{
Returns jobList.
}
\seealso{
Other f2dbBatch: 
\code{\link{f2dbBatch}},
\code{\link{f2dbBatch-class}},
\code{\link{f2dbRun,f2dbBatch-method}},
\code{\link{jobList-set-method}}
}
\concept{f2dbBatch}
