% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-f2dbBatch.R
\docType{methods}
\name{jobList-set-method}
\alias{jobList-set-method}
\alias{jobList<-}
\alias{jobList<-,f2dbBatch,f2dbJob-method}
\alias{jobList<-,f2dbBatch,list-method}
\title{jobList<-}
\usage{
jobList(object) <- value

\S4method{jobList}{f2dbBatch,f2dbJob}(object) <- value

\S4method{jobList}{f2dbBatch,list}(object) <- value
}
\arguments{
\item{object}{An \code{f2dbBatch}.}

\item{value}{An \code{f2dbJob}.}
}
\description{
Add \code{f2dbJob} objects to the end of the job list.
}
\seealso{
Other f2dbBatch: 
\code{\link{f2dbBatch}},
\code{\link{f2dbBatch-class}},
\code{\link{jobList-method}}
}
\concept{f2dbBatch}
