% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R, R/show-f2dbJob.R
\docType{methods}
\name{listTasks-method}
\alias{listTasks-method}
\alias{listTasks}
\alias{listTasks,f2dbJob-method}
\title{listTasks}
\usage{
listTasks(object)

\S4method{listTasks}{f2dbJob}(object)
}
\arguments{
\item{object}{An \code{f2dbJob} object.}
}
\value{
A list of named character vectors describing each task.
}
\description{
Shows job tasks.
}
\seealso{
Other f2dbJob: 
\code{\link{addNewTask-method}},
\code{\link{f2dbJob}},
\code{\link{f2dbJob-class}},
\code{\link{jobInput-set-method}}
}
\concept{f2dbJob}
