% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-f2dbObject.R, R/name-f2dbObject.R
\docType{methods}
\name{name-method}
\alias{name-method}
\alias{name}
\alias{name,f2dbObject-method}
\title{name}
\usage{
name(object)

\S4method{name}{f2dbObject}(object)
}
\arguments{
\item{object}{An \code{f2dbObject}.}
}
\value{
A name.
}
\description{
Returns the name of an \code{f2dbObject}.
}
\seealso{
Other f2dbObject: 
\code{\link{f2dbObject}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbObject}
