% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R
\docType{methods}
\name{nextTask-method}
\alias{nextTask-method}
\alias{nextTask}
\alias{nextTask,f2dbTask-method}
\title{nextTask}
\usage{
nextTask(object)

\S4method{nextTask}{f2dbTask}(object)
}
\arguments{
\item{object}{An \code{f2dbTask}}
}
\value{
An \code{f2dbTask}
}
\description{
Returns the \code{nextTask}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbListTask-class}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}
}
\concept{f2dbTask}
