% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R, R/f2dbEndTask.R
\docType{methods}
\name{nextTask-set-method}
\alias{nextTask-set-method}
\alias{nextTask<-}
\alias{nextTask<-,f2dbTask,f2dbTask-method}
\alias{nextTask<-,f2dbEndTask,f2dbTask-method}
\title{nextTask<-}
\usage{
nextTask(object) <- value

\S4method{nextTask}{f2dbTask,f2dbTask}(object) <- value

\S4method{nextTask}{f2dbEndTask,f2dbTask}(object) <- value
}
\arguments{
\item{object}{An \code{f2dbTask}.}

\item{value}{An \code{f2dbTask}.}
}
\description{
Sets the \code{nextTask}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbAppendTable}()},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}
}
\concept{f2dbTask}
