% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_conn.R
\name{set_dbconn}
\alias{set_dbconn}
\title{Set database connection}
\usage{
set_dbconn(conn)
}
\arguments{
\item{conn}{A DBIConnection object returned by a call to dbConnect.}
}
\value{
\code{conn}, invisibly.
}
\description{
Sets the database connection to be used when loading data.
}
\details{
The user is responsible for both opening and closing the database connection.
}
