% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{setf2dbLogger}
\alias{setf2dbLogger}
\title{setf2dbLogger}
\usage{
setf2dbLogger(logger)
}
\arguments{
\item{logger}{An object of \code{log4r::logger} class.}
}
\description{
Sets the file2db logger.
}
\details{
\code{file2db} uses the package \code{log4r}. The default log is
\code{log4r::logger("INFO", appenders = log4r::file_appender("f2db.log"))}.
}
\concept{logging}
