% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{setTaskLogger}
\alias{setTaskLogger}
\title{setTaskLogger}
\usage{
setTaskLogger(logger)
}
\description{
Sets the task logger.
}
\details{
\code{file2db} uses the package \code{log4r} to maintain two logs:
\itemize{
\item \code{f2db.log} documents the process of running an \code{f2dbObject}; e.g., name,
number of tasks, task function, etc. The default \code{f2db.log} is
\code{log4r::logger("INFO", appenders = log4r::file_appender("f2db.log"))}.
\item \code{task.log} logs the results of executing a task function. The default
\code{task.log} is
\code{log4r::logger("INFO", appenders = log4r::file_appender("task.log"))}
}

Use \code{setf2dbLogger} and \code{setTaskLogger} to modify the default loggers.
}
\seealso{
Other logging: 
\code{\link{setf2dbLogger}()}
}
\concept{logging}
