% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-f2dbTaskFunction.R
\docType{methods}
\name{taskCall-method}
\alias{taskCall-method}
\alias{taskCall}
\alias{taskCall,f2dbTaskFunction-method}
\title{taskCall}
\usage{
taskCall(object)

\S4method{taskCall}{f2dbTaskFunction}(object)
}
\arguments{
\item{object}{An \code{f2dbTaskFunction} object.}
}
\value{
The \code{taskCall} evaluated when the given object is run.
}
\description{
Returns the task call of an \code{f2dbTaskFunction}.
}
\seealso{
Other f2dbTaskFunction: 
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2dbTaskFunction}
