% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R
\docType{methods}
\name{taskFunction-method}
\alias{taskFunction-method}
\alias{taskFunction}
\alias{taskFunction,f2dbTask-method}
\title{taskFunction}
\usage{
taskFunction(object)

\S4method{taskFunction}{f2dbTask}(object)
}
\arguments{
\item{object}{An \code{f2dbTask}}
}
\value{
An \code{f2dbTaskFunction}
}
\description{
Returns the \code{f2dbTaskFunction} held by the \code{f2dbTask}
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-set-method}}
}
\concept{f2dbTask}
