% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R
\docType{methods}
\name{taskFunction-set-method}
\alias{taskFunction-set-method}
\alias{taskFunction<-}
\alias{taskFunction<-,f2dbTask,f2dbTaskFunction-method}
\title{taskFunction<-}
\usage{
taskFunction(object) <- value

\S4method{taskFunction}{f2dbTask,f2dbTaskFunction}(object) <- value
}
\arguments{
\item{object}{An \code{f2dbTask}.}

\item{value}{An \code{f2dbTaskFunction}.}
}
\description{
Sets the \code{f2dbTaskFunction} held by the \code{f2dbTask}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask}},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{name-method}},
\code{\link{taskFunction-method}}
}
\concept{f2dbTask}
