% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R
\docType{methods}
\name{taskList-method}
\alias{taskList-method}
\alias{taskList}
\alias{taskList,f2dbJob-method}
\title{taskList}
\usage{
taskList(object)

\S4method{taskList}{f2dbJob}(object)
}
\arguments{
\item{object}{An \code{f2dbJob} object.}
}
\value{
A list of \code{f2dbTask}s.
}
\description{
Returns task List
}
\seealso{
Other f2dbJob: 
\code{\link{f2dbJob}},
\code{\link{f2dbJob-class}},
\code{\link{jobInput-set-method}},
\code{\link{taskList-set-method}}
}
\concept{f2dbJob}
