% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R
\docType{methods}
\name{taskList-set-method}
\alias{taskList-set-method}
\alias{taskList<-}
\alias{taskList<-,f2dbJob,f2dbTask-method}
\alias{taskList<-,f2dbJob,list-method}
\alias{taskList<-,f2dbJob,NULL-method}
\title{taskList<-}
\usage{
taskList(object) <- value

\S4method{taskList}{f2dbJob,f2dbTask}(object) <- value

\S4method{taskList}{f2dbJob,list}(object) <- value

\S4method{taskList}{f2dbJob,NULL}(object) <- value
}
\arguments{
\item{object}{An \code{f2dbJob}.}

\item{value}{An \code{f2dbTask}.}
}
\description{
Add \code{f2dbTask} objects to the end of the task list and update the \code{nextTask}
slot of the next-to-last task to point to the newly added task.
}
\seealso{
Other f2dbJob: 
\code{\link{f2dbJob}},
\code{\link{f2dbJob-class}},
\code{\link{jobInput-set-method}},
\code{\link{taskList-method}}
}
\concept{f2dbJob}
