% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2dbEndTask.R
\docType{class}
\name{f2dbEndTask-class}
\alias{f2dbEndTask-class}
\title{f2dbEndTask class}
\description{
A do-nothing task. This is the default nextTask for f2dbTask instances.
}
\section{Slots}{

\describe{
\item{\code{name}}{"ENDTASK"}
}}

\seealso{
Other file2db classes: 
\code{\link{f2dbObject.class}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask.class}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}
}
\concept{file2db classes}
