% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2dbEndTask.R
\name{f2dbEndTask}
\alias{f2dbEndTask}
\title{f2dbEndTask}
\usage{
f2dbEndTask()
}
\value{
An \code{f2dbEndTask} object.
}
\description{
\code{f2dbEndTask} constructor.
}
\seealso{
Other f2dbEndTask: 
\code{\link{f2dbEndTask-class}}

Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}}
}
\concept{f2dbEndTask}
\concept{f2dbTask}
