% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-f2dbJob.R
\docType{class}
\name{f2dbJob-class}
\alias{f2dbJob-class}
\title{f2dbJob class}
\description{
A container for job tasks.
}
\section{Slots}{

\describe{
\item{\code{name}}{Job name. This should be unique within the job batch.}

\item{\code{jobInput}}{Input supplied to the first task.}

\item{\code{tasks}}{List of job tasks.}
}}

\seealso{
Other f2dbJob: 
\code{\link{appendTask-method}},
\code{\link{f2dbJob}},
\code{\link{jobInput-set-method}}

Other f2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbJob}
