% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-file2dbObject.R
\docType{class}
\name{f2dbObject-class}
\alias{f2dbObject-class}
\title{f2dbObject class}
\description{
Base class from which all other f2db classes descend. This is a
virtual class.
}
\seealso{
Other file2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbTask-class}},
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction.class}}
}
\concept{file2db classes}
