% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbEndTask.R
\docType{methods}
\name{f2dbRun,f2dbEndTask-method}
\alias{f2dbRun,f2dbEndTask-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbEndTask}(object, ...)
}
\arguments{
\item{object}{An \code{f2dbEndTask} to run.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Unused.}
}
\value{
A list:
\item{success}{\code{TRUE}.}
\item{object}{The type of object being run.}
\item{name}{The name of the object being run.}
\item{messages}{Character vector "".}
}
\description{
Run an \code{f2dbEndTask}.
}
\details{
\code{f2dbEndTask}s mark then end of a task list and will always return \code{TRUE}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbRun methods}
\concept{f2dbTask}
