% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbTask.R
\docType{methods}
\name{f2dbRun,f2dbTask-method}
\alias{f2dbRun,f2dbTask-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbTask}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbTask} to run.}

\item{input}{Input to pass to the underlying task function.}

\item{item}{The job item being processed.}
}
\value{
Logical value indicating success (\code{TRUE}) or failure
(\code{FALSE}). \code{TRUE} is returned if both \code{taskFunction} and \code{nextTask} are
successful. Otherwise, \code{FALSE} is returned.
}
\description{
Run an \code{f2dbTask}.
}
\details{
When called on an \code{f2dbTask} object, \code{f2dbRun} will run \code{taskFunction}. If
\code{taskFunction} is successful, then \code{f2dbRun} is called with \code{nextTask}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbListTask-class}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbListTask-method}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbRun methods}
\concept{f2dbTask}
