% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R, R/f2dbEndTask.R
\docType{methods}
\name{f2dbRun,f2dbTask-method}
\alias{f2dbRun,f2dbTask-method}
\alias{f2dbRun,f2dbEndTask-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbTask}(object, input = NA, item = NA)

\S4method{f2dbRun}{f2dbEndTask}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbTask} object.}

\item{input}{Task input.}

\item{item}{The current item being processed.}
}
\value{
A list:
\item{success}{Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
Success only indicates that it is safe to proceed; it does not mean there
are no warnings or errors.}
\item{output}{Run results, if any.}
}
\description{
Executes the given task and passes output to the next task.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask}},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{name-method}},
\code{\link{nextTask-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbRun methods}
\concept{f2dbTask}
