% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-f2dbTaskFunction.R
\docType{methods}
\name{f2dbRun,f2dbTaskFunction-method}
\alias{f2dbRun,f2dbTaskFunction-method}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbTaskFunction}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbTaskFunction} to run.}

\item{input}{Input to pass to the underlying task function.}

\item{item}{The job item being processed. This is passed to the task function
if the \code{f2dbTaskFunction} was created with an \code{itemName}.}
}
\value{
A list:
\item{success}{Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
Success only indicates that it is safe to proceed; it does not mean there
are no warnings or messages.}
\item{output}{Task function output}
\item{cnds}{Messages generated by the task function.}
}
\description{
Run an \code{f2dbTaskFunction}.
}
\details{
\code{f2dbRun} only returns \code{FALSE} if an error is thrown.
}
\seealso{
Other f2dbTaskFunction: 
\code{\link{f2dbTaskFunction}},
\code{\link{f2dbTaskFunction-class}},
\code{\link{taskCall-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbBatch-method}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbIterateTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun-method}}
}
\concept{f2dbRun methods}
\concept{f2dbTaskFunction}
