% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-f2dbObject.R, R/run-f2dbEndTask.R,
%   R/run-f2dbObject.R, R/run-f2dbTask.R, R/run-f2dbTaskFunction.R
\docType{methods}
\name{f2dbRun-method}
\alias{f2dbRun-method}
\alias{f2dbRun}
\alias{f2dbRun,f2dbEndTask-method}
\alias{f2dbRun,f2dbObject-method}
\alias{f2dbRun,f2dbTask-method}
\alias{f2dbRun,f2dbTaskFunction-method}
\title{f2dbRun}
\usage{
f2dbRun(object, ...)

\S4method{f2dbRun}{f2dbEndTask}(object, input = NA, item = NA)

\S4method{f2dbRun}{f2dbObject}(object, ...)

\S4method{f2dbRun}{f2dbTask}(object, input = NA, item = NA)

\S4method{f2dbRun}{f2dbTaskFunction}(object, input = NA, item = NA)
}
\arguments{
\item{object}{An \code{f2dbObject} to run}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Parameters passed to
class-specific \code{f2dbRun} implementations.}

\item{input}{Input passed to the object being run.}

\item{item}{Batch job}
}
\value{
A list:
\item{success}{Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
Success only indicates that it is safe to proceed; it does not mean there
are no warnings or errors.}
\item{...}{Additional values, depending upon the object that is run.}
}
\description{
Run an \code{f2dbObject}.
}
\seealso{
Other f2dbObject: 
\code{\link{f2dbObject}},
\code{\link{f2dbObject-class}},
\code{\link{name-method}}
}
\concept{f2dbObject}
\concept{f2dbRun methods}
