% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-f2dbObject.R
\docType{methods}
\name{f2dbRun-method}
\alias{f2dbRun-method}
\alias{f2dbRun}
\title{f2dbRun}
\usage{
f2dbRun(object, ...)
}
\arguments{
\item{object}{An \code{f2dbObject} to run}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Parameters passed to
class-specific \code{f2dbRun} implementations.}
}
\value{
All objects return a list with the following values:
\item{success}{Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
Success only indicates that it is safe to proceed; it does not mean there
are no warnings or errors.}
\item{object}{The type of object being run.}
\item{name}{The name of the object being run.}

Additional values are returned depending on the type of object being run. See
help topics for object-specific \code{f2dbRun} implementations.
}
\description{
Run an \code{f2dbObject}.
}
\seealso{
Other f2dbObject: 
\code{\link{f2dbObject}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{name-method}}

Other f2dbRun methods: 
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbObject-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbRun,f2dbTaskFunction-method}}
}
\concept{f2dbObject}
\concept{f2dbRun methods}
