% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-file2dbTaskFunction.R
\docType{methods}
\name{f2dbRun.f2dbTaskFunction}
\alias{f2dbRun.f2dbTaskFunction}
\title{f2dbRun}
\usage{
\S4method{f2dbRun}{f2dbTaskFunction}(.Object, input = NA, checklistItem = NA)
}
\arguments{
\item{object}{An \code{f2dbTaskFunction} object}
}
\value{
Logical value indicating success (\code{TRUE}) or failure (\code{FALSE}).
Success only indicates that it is safe to proceed; it does not mean there are
no warnings or errors.
}
\description{
Runs the given task function.
}
\seealso{
Other f2dbRun methods: 
\code{\link{f2dbRun-generic}}
}
\concept{f2dbRun methods}
