% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-f2dbTask.R
\docType{class}
\name{f2dbTask-class}
\alias{f2dbTask-class}
\title{f2dbTask class}
\description{
A class for encapsulating an individual task.
}
\section{Slots}{

\describe{
\item{\code{name}}{Task name. This should be unique within the job.}

\item{\code{taskFunction}}{An \code{f2dbTaskFunction} object to perform the actual work of
the task.}

\item{\code{nextTask}}{An \code{f2dbTaskObject} object that performs the next task in the
job sequence. \code{nextTask} takes the output of the current task as its input.}
}}

\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbListTask-class}},
\code{\link{f2dbRun,f2dbEndTask-method}},
\code{\link{f2dbRun,f2dbJob-method}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{nextTask-method}},
\code{\link{nextTask-set-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}

Other f2db classes: 
\code{\link{f2dbBatch-class}},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbJob-class}},
\code{\link{f2dbListTask-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTaskFunction-class}}
}
\concept{f2db classes}
\concept{f2dbTask}
