% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-file2dbTaskFunction.R
\docType{class}
\name{f2dbTaskFunction-class}
\alias{f2dbTaskFunction-class}
\title{f2dbTaskFunction class}
\description{
The f2dbTaskFunction class encapsulates the function that accomplishes the
actual work of the task. Along with the \code{f2dbRun} method, this class is
responsible for:
\enumerate{
\item Calling the underlying task function
\item Capturing and logging any warnings or errors
\item Indicating whether or not the task executed successfully
\item Returning task function output
}
}
\details{
\strong{must} implement a method for the generic \code{run()}
}
\section{Slots}{

\describe{
\item{\code{taskFunction}}{The task function}

\item{\code{auxParameters}}{Additional parameters to be passed to the task function}
}}

\seealso{
Other file2db classes: 
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}}
}
\concept{file2db classes}
