% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-f2dbTaskFunction.R
\docType{class}
\name{f2dbTaskFunction-class}
\alias{f2dbTaskFunction-class}
\title{f2dbTaskFunction class}
\description{
The \code{f2dbTaskFunction} class encapsulates the function that accomplishes the
actual work of the task. Along with \code{\link[=f2dbRun]{f2dbRun()}} method, this class is
responsible for:
\enumerate{
\item Calling the underlying task function.
\item Capturing any warnings or errors.
\item Returning task function output.
\item Returning whether or not the task executed successfully.
}
}
\details{
Classes inheriting from \code{f2dbTaskFunction} should fulfill the
responsibilities listed above and implement the \code{\link[=f2dbRun]{f2dbRun()}} generic method.
}
\section{Slots}{

\describe{
\item{\code{name}}{Set to the name of the underlying task function by the constructor.}

\item{\code{taskCall}}{A call to the underlying task function.}
}}

\seealso{
Other f2dbTaskFunction: 
\code{\link{f2dbRun,f2dbTaskFunction-method}},
\code{\link{f2dbTaskFunction}},
\code{\link{taskCall-method}}

Other f2db classes: 
\code{\link{f2dbBatch-class}},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbIterateTask-class}},
\code{\link{f2dbJob-class}},
\code{\link{f2dbObject-class}},
\code{\link{f2dbTask-class}}
}
\concept{f2db classes}
\concept{f2dbTaskFunction}
