% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-f2dbTaskFunction.R
\name{f2dbTaskFunction}
\alias{f2dbTaskFunction}
\title{f2dbTaskFunction}
\usage{
f2dbTaskFunction(taskFunction, ..., inputName, itemName)
}
\arguments{
\item{taskFunction}{Function to call when running the task.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Parameters passed to the task
function.}

\item{inputName}{By default, task input is passed to the task function as an
unnamed first parameter. If \code{inputName} is provided, task input will be
passed using \code{inputName}.}

\item{itemName}{By default, the current batch item is not passed to the task
function. If \code{itemName} is provided, the batch item will be passed using
\code{itemName}.}
}
\value{
An \code{f2dbTaskFunction} object.
}
\description{
\code{f2dbTaskFunction} constructor.
}
\seealso{
Other f2dbTaskFunction: 
\code{\link{f2dbTaskFunction-class}},
\code{\link{taskCall-method}}
}
\concept{f2dbTaskFunction}
