% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2dbEndTask.R
\name{initialize,f2dbEndTask-method}
\alias{initialize,f2dbEndTask-method}
\title{Initialize function for f2dbEndTask objects}
\usage{
\S4method{initialize}{f2dbEndTask}(.Object)
}
\description{
This initializer explicitly avoids calling the parent f2dbTask initializer.
}
