% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-f2dbObject.R, R/02-f2dbTask.R
\docType{methods}
\name{name-method}
\alias{name-method}
\alias{name}
\alias{name,f2dbObject-method}
\alias{name,f2dbTask-method}
\title{name}
\usage{
name(object)

\S4method{name}{f2dbObject}(object)

\S4method{name}{f2dbTask}(object)
}
\arguments{
\item{object}{An \code{f2dbObject}.}
}
\value{
A name.
}
\description{
Returns the name of an \code{f2dbObject}.
}
\seealso{
Other f2dbTask: 
\code{\link{f2dbEndTask}},
\code{\link{f2dbEndTask-class}},
\code{\link{f2dbRun,f2dbTask-method}},
\code{\link{f2dbTask}},
\code{\link{f2dbTask-class}},
\code{\link{nextTask-method}},
\code{\link{taskFunction-method}},
\code{\link{taskFunction-set-method}}
}
\concept{f2dbTask}
